.history/requirements_20260107101530.txt
flask
flask-cors
tensorflow
torch
transformers
scikit-learn
pandas
numpy
librosa
pymongo
dnspython
soundfile
pyaudio
openSMILE
sentencepiece
pydub
matplotlib
speechrecognition
pyttsx3
python-dotenv
